+++ crates/utils/Cargo.toml
@@ -12,7 +12,7 @@ all-features = true
 [dependencies]
 starbase_styles = { version = "0.1.18", path = "../styles" }
 dirs = "5.0.1"
-miette = { workspace = true }
+miette = { workspace = true, optional = true }
 once_cell = { workspace = true }
 relative-path = { workspace = true }
 regex = { workspace = true, optional = true }
@@ -45,6 +45,7 @@ default = []
 editor-config = ["dep:ec4rs"]
 fs-lock = ["dep:fs4", "dep:tokio"]
 glob = ["dep:wax"]
+miette = ["dep:miette"]
 json = ["dep:json_comments", "dep:regex", "dep:serde", "dep:serde_json"]
 toml = ["dep:toml", "dep:serde"]
 yaml = ["dep:regex", "dep:serde", "dep:serde_yaml"]
