+++ crates/utils/src/toml.rs
@@ -1,47 +1,15 @@
-use crate::fs::{self, FsError};
-use miette::Diagnostic;
+use crate::fs;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
-use starbase_styles::{Style, Stylize};
-use std::path::{Path, PathBuf};
-use thiserror::Error;
+use std::path::Path;
 use tracing::trace;
 
+pub use crate::toml_error::TomlError;
 pub use toml as serde_toml;
 pub use toml::{
     from_str, to_string, to_string_pretty,
     value::{Datetime as TomlDatetime, Table as TomlTable, Value as TomlValue},
 };
-
-#[derive(Error, Diagnostic, Debug)]
-pub enum TomlError {
-    #[diagnostic(transparent)]
-    #[error(transparent)]
-    Fs(#[from] FsError),
-
-    #[diagnostic(code(toml::parse_file))]
-    #[error("Failed to parse TOML file {}.", .path.style(Style::Path))]
-    ReadFile {
-        path: PathBuf,
-        #[source]
-        error: toml::de::Error,
-    },
-
-    #[diagnostic(code(toml::stringify))]
-    #[error("Failed to stringify TOML.")]
-    Stringify {
-        #[source]
-        error: toml::ser::Error,
-    },
-
-    #[diagnostic(code(toml::stringify_file))]
-    #[error("Failed to stringify TOML for file {}.", .path.style(Style::Path))]
-    StringifyFile {
-        path: PathBuf,
-        #[source]
-        error: toml::ser::Error,
-    },
-}
 
 /// Read a file at the provided path and deserialize into the required type.
 /// The path must already exist.
